core v.2.1
Убрал лишний тест
